requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-core
langchain-ollama
Flask
line-bot-sdk